AgggregateCounterVector
Like aggregatecounter but relies on a vector, where each idx is
guaranteed to be per-thread

Also gives an idea of sub-count per buckets using a kinde of bloom
filter. Eg if the mask is 512-1, will give an idea of how many of full
aggregate per agg & mask there is.

The use is for RSS bucket counting. It will count the number of packets
per RSS bucket, and an idea of how many flows.