Rework RSS++ code into a library + integration
We added a NICScheduler library that implements all the logic of RSS++.
DeviceBalancer is only a mere integration of the library.

The result is that someone could reuse the scheduling library (and
therefore the RSS++ scheduler, the Metron scheduler, ...) much more
easily. Only a few dependencies are left to remove.